Volcano Application
This is a robot programme that can determine how hot the temperature is
by having the robot in a sub class that uses the data methods in the
super class. This type of programming is useful, because you can have as
many robot vehicles as you like and you don't need to write the same
code over and over again because it extracts the methods from the above
super class.